okay full algo finished wrote tests for it now seems to work. but when i use optimisation flags the program seems to get fucked

+++ inc/distance.h
@@ -2,8 +2,8 @@
 #define VT_DISTANCE_H
 typedef struct vector vec;
 
-typedef float (*distance_func) (vec* a, vec* b);
+typedef long double (*distance_func) (vec* a, vec* b);
 
- float l2_sq_distance(vec* a, vec* b);
+ long double l2_sq_distance(vec* a, vec* b);
 
 #endif
